removed component re-exports and renamed useLocalesMap filename
