--- components/icons/index.js
@@ -1,8 +1,0 @@
-export { default as BackendAgnosticIcon } from "./backend-agnostic";
-export { default as LightweightIcon } from "./lightweight";
-export { default as PaginationIcon } from "./pagination";
-export { default as RealtimeIcon } from "./realtime";
-export { default as RemoteLocalIcon } from "./remote-local";
-export { default as RenderingStrategiesIcon } from "./rendering-strategies";
-export { default as SuspenseIcon } from "./suspense";
-export { default as TypeScriptIcon } from "./typescript";
