--- components/index.js
@@ -1,7 +1,0 @@
-export { default as Authors, Author } from "./authors";
-export { default as BlogIndex } from "./blog-index";
-export { default as Features, Feature } from "./features";
-export { default as Logo } from "./logo";
-export { default as Vercel } from "./vercel";
-export { default as Video } from "./video";
-export { default as useLocalesMap } from "./useLocalesMap";
